Swap `unset_args` and `unknown_args` in the return of the `set_arguments`
Swap the values since it's more probable to get a nil `unset_args` than
a nil `unknown_args`, because an `unknown_args` may not be an error.

Since there can be two sets of flags (one for the whole program and
another for the subcommand), an unknown argument for a set may not be an
unknown argument for the other.